+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.2] - 2024-04-29
+### Anchor Around UI Object
+- Added editor script to anchor UI Object perfectly around it
+- Default Keybind is CTRL + SHIFT + Q
+
 ## [1.0.1] - 2024-04-29
 ### Slider support
 - Added Slider support for PriosEventReciever